First model
Honestly we might be done at this point, the r squared value for this model is over 97%, so we might just be done here. I recommend maybe messing around with stuff in a new branch to see if you maybe catch something or find this model overfits the train data.